refactor: remove deprecated env
