--- matd3/environments/myenv/make_env.py
@@ -1,6 +1,0 @@
-from matd3.environments.myenv.env_new import MyEnv
-
-
-def make_env():
-    # todo：补全，传入必要参数去初始化训练场景
-    return MyEnv()
